feat: Add toggle completed task functionality
This commit adds the ability to toggle the completed status of a task. The `TOGGLE_COMPLETED_TASK` constant has been added to the `apiConfig.js` file to represent the API endpoint for toggling the completed status. Additionally, the `Tasks` component has been imported and rendered in the `processPagePJESeabra.vue` and `processPagePJETRT.vue` files to display the tasks associated with a process.